+++ db/migrate/20150403212139_fix_wrong_only_sharing.rb
@@ -1,0 +1,11 @@
+class FixWrongOnlySharing < ActiveRecord::Migration
+  def up
+    Contact.where(sharing: true, receiving: false)
+      .where(id: AspectMembership.select(:contact_id))
+      .update_all(receiving: true)
+  end
+
+  def down
+    raise ActiveRecord::IrreversibleMigration
+  end
+end